requirements.txt
Django~=4.2.4
django-debug-toolbar==3.2.4
Pillow==10.0.0
pytz==2023.3
sorl-thumbnail==12.9.0
sqlparse==0.4.4

djangorestframework~=3.14.0